+++ src/Entelect.Encentivize.Sdk/Points/MemberAdhocInput.cs
@@ -2,7 +2,7 @@
 {
     public class MemberAdhocInput
     {
-        public int PointsToAdd { get; set; }
+        public decimal PointsToAdd { get; set; }
         public string DisplayComment { get; set; }
         public string AdditionalInformation { get; set; }
     }
